♻️ Rename existing router + controllers as v1
To allow the PUT action to replace only parts of the record, we need to
implement a v2 of the API so that we maintain backwards compatibility
with the already deployed applications. This moves all the existing
endpoints in to a v1 router in preparation of that.

Issue: Scottish-Natural-Heritage/Licensing#577